+++ chartlets.py/demo/my_extension/my_panel_8.py
@@ -5,8 +5,15 @@
 import altair as alt
 
 from chartlets import Component, Input, State, Output
-from chartlets.components import (Tabs, Tab, Typography, Box,
-                                  VegaChart, Table)
+from chartlets.components import (
+    Tabs,
+    Tab,
+    Typography,
+    Box,
+    VegaChart,
+    Table,
+    Button,
+)
 from chartlets.components.table import TableColumn, TableRow
 
 from server.context import Context
@@ -41,25 +48,48 @@ def render_panel(
         ["3", "Peter", "Jones", 40],
     ]
 
-    table = Table(id="table", rows=rows, columns=columns, hover=True)
-
-    info_text = Typography(id="info_text", children=["This is a text."])
-    chart = VegaChart(
-        id="chart", chart=(
-        alt.Chart(dataset)
-        .mark_bar()
-        .encode(
-            x=alt.X("x:N", title="x"),
-            y=alt.Y("a:Q", title="a"))
-        .properties(width=290, height=300, title="Vega charts")
-    ), style={"flexGrow": 1}
+    open_button = Button(
+        id="open_button", text="Show component!", style={"margin": "20px"}
     )
 
-    tab1 = Tab(id = "tab1", label="Tab 1", children=[table])
-    tab2 = Tab(id = "tab2", label="Tab 2", children=[info_text])
-    tab3 = Tab(id="tab3", label="Tab 3", children=[chart])
+    table = Table(
+        id="table",
+        rows=rows,
+        columns=columns,
+        hover=True,
+        style={"width": "100%", "padding": "2px"},
+    )
 
-    tabs = Tabs(id = "tabs", value = 0, children=[tab1,tab2,tab3])
+    chart = VegaChart(
+        id="chart",
+        chart=(
+            alt.Chart(dataset)
+            .mark_bar()
+            .encode(x=alt.X("x:N", title="x"), y=alt.Y("a:Q", title="a"))
+            .properties(width=290, height=300, title="Vega charts")
+        ),
+        style={"flexGrow": 1, "margin": "10px"},
+    )
+
+    tab1 = Tab(
+        id="tab1",
+        label="Tab 1",
+        children=[table],
+        style={"backgroundColor": "darkblue", "padding": "1px"},
+    )
+    tab2 = Tab(id="tab2", label="Tab 2", children=["This is a text."], disabled=True)
+    tab3 = Tab(
+        id="tab3",
+        label="Tab 3",
+        children=[chart],
+    )
+
+    tabs = Tabs(
+        id="tabs",
+        value=0,
+        children=[tab1, tab2, tab3],
+        style={"visibility": "hidden"},
+    )
 
     return Box(
         style={
@@ -68,6 +98,27 @@ def render_panel(
             "width": "100%",
             "height": "100%",
         },
-        children=[ tabs ],
+        children=[open_button, tabs],
     )
 
+
+# noinspection PyUnusedLocal
+@panel.callback(
+    Input("open_button", "clicked"),
+    Input("tabs", "style"),
+    Output("tabs", "style"),
+    Output("open_button", "text"),
+)
+def tabs_on_open(ctx: Context, button, style) -> tuple[dict, str]:
+    visibility = style["visibility"]
+
+    if visibility == "hidden":
+        return (
+            {**style, "visibility": "visible"},
+            "Hide component!",
+        )
+    else:
+        return (
+            {**style, "visibility": "hidden"},
+            "Show component!",
+        )